+++ src/client/ui/pages/not_found.rs
@@ -26,8 +26,8 @@ impl UiPageInterface<Config> for Error404Page {
             .stroke(Stroke::NONE)
             .shadow(Shadow {
                 color: style.visuals.window_shadow.color,
-                spread: style.spacing.item_spacing.y / 2.0,
-                blur: 5.0,
+                spread: (style.spacing.item_spacing.y / 2.0) as u8,
+                blur: 5,
                 ..Default::default()
             })
             .show(ui, |ui| {
